fix(security): 修正 JWT 令牌解析中的用户 ID 类型，从 Integer 更改为 Long，避免溢出错误

+++ server/src/main/java/com/tomato/bookstore/security/JwtTokenProvider.java
@@ -139,6 +139,6 @@ public class JwtTokenProvider {
      * @return 用户 ID
      */
     public Long extractUserId(String token) {
-        return Long.valueOf(extractClaim(token, claims -> claims.get("userId", Integer.class)).longValue());
+        return Long.valueOf(extractClaim(token, claims -> claims.get("userId", Long.class)).longValue());
     }
 }
